Did some improvements to the movie scanner
