+++ server/cmd/api/movies_scanner_helpers.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"database/sql"
 	"fmt"
 	"igloo/cmd/internal/database"
 	"igloo/cmd/internal/helpers"
@@ -24,7 +23,6 @@ func extractYearFromReleaseDate(releaseDate string) int {
 func (app *Application) getOrCreateArtist(
 	ctx context.Context,
 	qtx *database.Queries,
-	scan *movieScanContext,
 	tmdbID int,
 	name string,
 	profilePath string,
@@ -40,39 +38,5 @@ func (app *Application) getOrCreateArtist(
 		return nil, fmt.Errorf("upsert artist failed: %w", err)
 	}
 
-	if scan != nil {
-		scan.artistIDs[tmdbID] = upserted.ID
-	}
-
 	return &upserted, nil
-}
-
-// Savepoints let one scanner item fail without rolling back its batch.
-func manageSavepoint(
-	ctx context.Context,
-	tx *sql.Tx,
-	savepointName string,
-	fn func() error,
-) error {
-	_, err := tx.ExecContext(ctx, fmt.Sprintf("SAVEPOINT %s", savepointName))
-	if err != nil {
-		return fmt.Errorf("failed to create savepoint %s: %w", savepointName, err)
-	}
-
-	err = fn()
-	if err != nil {
-		_, rollbackErr := tx.ExecContext(ctx, fmt.Sprintf("ROLLBACK TO SAVEPOINT %s", savepointName))
-		if rollbackErr != nil {
-			return fmt.Errorf("failed to rollback savepoint %s (original error: %w): %w", savepointName, err, rollbackErr)
-		}
-
-		return err
-	}
-
-	_, err = tx.ExecContext(ctx, fmt.Sprintf("RELEASE SAVEPOINT %s", savepointName))
-	if err != nil {
-		return fmt.Errorf("failed to release savepoint %s: %w", savepointName, err)
-	}
-
-	return nil
 }
